New feature: External Endstop probing via G38 =============================================
G38 behaves similarly to G30, but instead of using the z-probe for sensing the contact, it uses an endstop (a digital input)

Currently the endstop that is selected is the "Secure_sw" endstop that can be found in the totumduino board, which corresponds to
pin NOT_SECURE_SW_PIN in the firmware. However, any other endstop may be selected by changing EXTERNAL_ENDSTOP_Z_PROBING_PIN in
Configuration.h.

Instead of an endstop switch, this can be configured to detect the mill touching a conductive surface (such as a copper clad to make a PCB).

From the hardware point of view, the two pins of connector "Secure_sw" are soldered to a 1Kohm resistor, each pin to one end of the resistor,
and the pin closest to the corner of the totumduino board is attached to the conductive surface (copper clad). I chose to solder a metal washer
to that pin, and attach the washer to the PCB using a screw and a nut, but any other attachment means would work.