Support for thermistor type 1
It should not be a problem to add type 1, as it requires not flash memory, because the table is already included for the heated bed which uses thermistor type 1.

If you do M802 you will get: ( 169, 11, 1 )

So type 1 is index 2 (0,1,2)

Change the maximum temperature using M801 (for example for 250ºC):

M801 S250

Enable the type 1 for the extruder using:

M800 S2

You can see if the setting was taken we doing afterwards M800, which will reply "2".